requirements.txt
aioredis==1.3.0
asgiref==3.2.2
asn1crypto==1.0.1
astroid==2.3.1
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==19.2.0
autobahn==19.10.1
Automat==0.7.0
cffi==1.12.3
channels==2.3.0
channels-redis==2.4.0
constantly==15.1.0
coverage==4.5.4
cryptography==2.7
daphne==2.3.0
Django==2.2.6
django-environ==0.4.5
djangorestframework==3.10.3
entrypoints==0.3
flake8==3.7.8
freezegun==0.3.12
gunicorn==19.9.0
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==0.23
incremental==17.5.0
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
more-itertools==7.2.0
msgpack==0.6.2
packaging==19.2
pluggy==0.13.0
psycopg2==2.8.3
py==1.8.0
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
PyHamcrest==1.9.0
PyJWT==1.7.1
pylint==2.4.2
pyparsing==2.4.2
pytest==5.2.0
pytest-cov==2.8.1
pytest-django==3.5.1
python-dateutil==2.8.0
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
Twisted==19.7.0
txaio==18.8.1
typed-ast==1.4.0
wcwidth==0.1.7
wrapt==1.11.2
zipp==0.6.0
zope.interface==4.6.0
